♻️ refactor: clients page
This commit refactors the clients page by using a custom hook from the users store to get the
clients data and replacing the previous way of filtering the clients. It also adds a search input to
filter the clients by name and displays a datalist with the options. When a client is selected, it
displays a QR code with the client's id.